+++ jobbergate-api/jobbergate_api/main.py
@@ -1,7 +1,6 @@
 """
 Main file to startup the fastapi server.
 """
-import sys
 from contextlib import asynccontextmanager
 
 import asyncpg
@@ -18,6 +17,7 @@ from jobbergate_api.apps.job_script_templates.routers import router as job_scrip
 from jobbergate_api.apps.job_scripts.routers import router as job_scripts_router
 from jobbergate_api.apps.job_submissions.routers import router as job_submissions_router
 from jobbergate_api.config import settings
+from jobbergate_api.logging import init_logging
 from jobbergate_api.storage import engine_factory, handle_fk_error
 
 subapp = FastAPI(
@@ -86,9 +86,7 @@ async def lifespan(_: FastAPI):
     This is the preferred method of handling lifespan events in FastAPI.
     For mor details, see: https://fastapi.tiangolo.com/advanced/events/
     """
-    logger.remove()
-    logger.add(sys.stderr, level=settings.LOG_LEVEL)
-    logger.info(f"Logging configured 📝 Level: {settings.LOG_LEVEL}")
+    init_logging()
 
     yield